MyApps-13: Move next page UseCase to submodules.
