+++ app/src/main/java/xyz/torquato/myapps/di/domain/DomainWebModule.kt
@@ -7,7 +7,6 @@ import dagger.hilt.components.SingletonComponent
 import xyz.torquato.myapps.domain.web.GetSelectedBookUseCase
 import xyz.torquato.myapps.domain.web.content.GetBooksUseCase
 import xyz.torquato.myapps.domain.web.content.GetCacheBooksUseCase
-import xyz.torquato.myapps.domain.web.pagging.GetNextPageUseCase
 import xyz.torquato.myaps.domain.api.pagging.IPagingRepository
 import xyz.torquato.myaps.domain.api.web.ILocalWebRepository
 import xyz.torquato.myaps.domain.api.web.IQueryRepository
@@ -50,14 +49,4 @@ object DomainWebModule {
     )
 
 
-
-    @Singleton
-    @Provides
-    fun provideGetNextPageUseCase(
-       repository: IPagingRepository,
-       queryRepository: IQueryRepository
-    ): GetNextPageUseCase = GetNextPageUseCase(
-        repository,
-        queryRepository
-    )
 }
