+++ domain/di/src/main/java/xyz/torquato/myapps/domain/di/UseCaseImpl.kt
@@ -4,8 +4,11 @@ import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import xyz.torquato.myaps.domain.api.pagging.IPagingRepository
 import xyz.torquato.myaps.domain.api.web.ILocalWebRepository
+import xyz.torquato.myaps.domain.api.web.IQueryRepository
 import xyz.torquato.myaps.domain.impl.web.content.SetSelectedBookUseCase
+import xyz.torquato.myaps.domain.impl.web.paging.GetNextPageUseCase
 import javax.inject.Singleton
 
 @Module
@@ -18,5 +21,15 @@ object UseCaseImpl {
         repository: ILocalWebRepository
     ): SetSelectedBookUseCase = SetSelectedBookUseCase(
         repository
+    )
+
+    @Singleton
+    @Provides
+    fun provideGetNextPageUseCase(
+        repository: IPagingRepository,
+        queryRepository: IQueryRepository
+    ): GetNextPageUseCase = GetNextPageUseCase(
+        repository,
+        queryRepository
     )
 }
